polls: Separate PollException, clarify code
