+++ src/HonzaBotner.Discord.Services/Commands/Polls/PollException.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace HonzaBotner.Discord.Services.Commands.Polls;
+
+public class PollException : Exception
+{
+    public PollException ()
+    {}
+
+    public PollException (string message)
+        : base(message)
+    {}
+
+    public PollException (string message, Exception innerException)
+        : base (message, innerException)
+    {}
+}
